libcamera: swstats_cpu: Optimize processYUV420Sharpness()
Optimize processYUV420Sharpness() by unroling the 2 inner loops
loping over the 3 rows x colums over the kernel.

Since the 4 corners of the kernel have a multiplication factor of 0
this reduces the 2 loops to just 5 additions (with 1 -4 multiplication).

This reduces the processFrame() processing time (with wantSharpness=true)
from ~15ms/frame to ~7ms/frame.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>